Provide user feedback (popups) for files

+++ client/src/main/java/client/elements/FileElement.java
@@ -1,12 +1,16 @@
 package client.elements;
 
 import client.scenes.MainCtrl;
+import client.scenes.NoteEditCtrl;
 import client.utils.DialogUtil;
 import client.utils.ServerUtils;
 import commons.FileEntity;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Hyperlink;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
@@ -16,20 +20,26 @@ import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.Optional;
+import java.util.ResourceBundle;
 
 public class FileElement extends BorderPane {
 
     private final MainCtrl mainCtrl;
+    private final NoteEditCtrl noteEditCtrl;
     private final ServerUtils server;
     private final DialogUtil dialogUtil;
+    private final ResourceBundle resourceBundle;
 
     private final FileEntity file;
 
-    public FileElement(MainCtrl mainCtrl, ServerUtils server, DialogUtil dialogUtil,
-                       FileEntity file) {
+    public FileElement(MainCtrl mainCtrl, NoteEditCtrl noteEditCtrl, ServerUtils server, DialogUtil dialogUtil,
+                       ResourceBundle resourceBundle, FileEntity file) {
         this.mainCtrl = mainCtrl;
+        this.noteEditCtrl = noteEditCtrl;
         this.server = server;
         this.dialogUtil = dialogUtil;
+        this.resourceBundle = resourceBundle;
         this.file = file;
 
         Hyperlink label = new Hyperlink(file.getName());
@@ -39,6 +49,7 @@ public class FileElement extends BorderPane {
             this.promptDownload();
         });
         label.setPadding(new Insets(0, 5, 0, 0));
+        label.setMaxWidth(200);
 
         Image deleteImage = new Image("appIcon/delete_icon.png");
         ImageView deleteIcon = new ImageView(deleteImage);
@@ -68,24 +79,34 @@ public class FileElement extends BorderPane {
         System.out.println("Downloading file " + file.getName() + " (" + file.getId() + ") to "
                 + saveAt.getAbsolutePath());
 
-        // TODO: provide user feedback
         try (FileOutputStream out = new FileOutputStream(saveAt);
              ByteArrayInputStream in = new ByteArrayInputStream(file.getData())) {
             in.transferTo(out);
+
+            dialogUtil.showDialog(this.resourceBundle, AlertType.INFORMATION,
+                    "popup.files.downloaded");
         } catch (IOException ex) {
             System.err.println("Failed to save note file to disk");
             ex.printStackTrace();
+
+            dialogUtil.showDialog(this.resourceBundle, AlertType.ERROR,
+                    "popup.files.downloadFailed");
         }
     }
 
     private void promptDelete() {
-        // TODO: prompt before deleting
-        this.delete();
+        Optional<ButtonType> response = dialogUtil.showDialog(this.resourceBundle, AlertType.CONFIRMATION,
+                "popup.files.confirmDelete");
+        if (response.isPresent() && response.get() == ButtonType.OK) {
+            this.delete();
+        }
     }
 
     private void delete() {
-        // TODO: provide user feedback
         server.deleteFile(file);
+        noteEditCtrl.refresh();
+        dialogUtil.showDialog(this.resourceBundle, AlertType.CONFIRMATION,
+                "popup.files.deleted");
     }
 
 }
